captech/requirements.txt
Django==2.2.1
djangorestframework==3.10.2
mysqlclient==1.4.4
Pillow==6.1.0
Tree==0.2.4
